requirements.txt
boto3
uvicorn
python-dotenv
pydantic
requests
pytz